Meat, HP, and MP are now all stored internally as longs, rather than as ints
git-svn-id: http://svn.code.sf.net/p/kolmafia/code@19629 b29ace70-8910-0410-8dcc-aa2fc6433167

+++ src/net/sourceforge/kolmafia/combat/Macrofier.java
@@ -600,7 +600,7 @@ public class Macrofier
 
 	public static void macroSkill( StringBuffer macro, int skillId )
 	{
-		int cost = SkillDatabase.getMPConsumptionById( skillId );
+		long cost = SkillDatabase.getMPConsumptionById( skillId );
 		if ( cost > KoLCharacter.getMaximumMP() )
 		{
 			return; // no point in even trying
